requirements.txt
# Django-specific packages
Django==1.5.1
South==0.7.6
django-browserid==0.7.1
django_compressor==1.4
django-model-utils==2.0.3
django-reversion==1.7
django-licensing==0.2.4
djangorestframework==2.3.13

pyelasticsearch==0.6
BeautifulSoup<4.0
lxml==3.0.2
cssselect==0.8
Pillow==2.3.0
psycopg2==2.4.6

isodate==0.4.9
pytz==2012h
refine-client==0.2.1
smartypants==1.6.0.3

ipython==0.13.1

webtest==2.0.14
django-webtest==1.7.6

# CENDARI EXTRA REQUIREMENTS
pycurl==7.19.3.1
#django-haystack==2.1.0
#django-haystack==1.2.7
#xapian-haystack==1.1.5beta
guess-language==0.2
## PIL==1.1.7
## vv - wjw - I had to specify this alternative URL for PIL - vv
## http://effbot.org/media/downloads/Imaging-1.1.7.tar.gz 
#tiffany==0.6.4
rq==0.3.13
django-rq==0.6.1
rdflib==4.1.0
django-widget-tweaks==1.3
parse==1.6.4
django-admin2==0.6.1
# Symlinks to system packages
# Symlink djadmin2
## symlink: xapian
## symlink: bsddb3

